api-gateway-project/gateway/requirements.txt
fastapi
uvicorn
httpx
pyjwt
pybreaker
pydantic